Move tests_memory to exampels
