+++ Cargo.toml
@@ -37,6 +37,10 @@ cc = "1.0"
 [dev-dependencies]
 criterion = "0.4.0"
 proptest = "0.9.1"
+jemallocator = "0.5.0"
+jemalloc-ctl = "0.5.0"
+lazy_static = "1.4.0"
+
 
 [[bench]]
 name = "bits_benchmark"
