+++ examples/check_real_memory.rs
@@ -5,16 +5,24 @@ use ckb_vm::{
     },
     run, Bytes, FlatMemory, SparseMemory, ISA_IMC,
 };
+use jemalloc_ctl::{epoch, stats};
+use jemallocator::Jemalloc;
 use lazy_static::lazy_static;
-use std::process::{id, Command};
+use std::{
+    process::{id, Command},
+    thread,
+};
 
 lazy_static! {
     pub static ref BIN_PATH_BUFFER: Bytes =
-        Bytes::from(&include_bytes!("../../programs/alloc_many")[..]);
+        Bytes::from(&include_bytes!("../tests/programs/alloc_many")[..]);
     pub static ref BIN_NAME: String = format!("alloc_many");
 }
 
-static G_CHECK_LOOP: usize = 100;
+#[global_allocator]
+static GLOBAL: Jemalloc = Jemalloc;
+
+static G_CHECK_LOOP: usize = 2;
 
 fn get_current_memory() -> usize {
     let pid = format!("{}", id());
@@ -36,7 +44,7 @@ fn get_current_memory() -> usize {
     memory_size.parse().unwrap()
 }
 
-fn check_int_real_memory(memory_size: usize) {
+fn check_interpreter(memory_size: usize) -> Result<(), ()> {
     println!(
         "Check interpreter memory overflow, ckb-vm memory size: {}",
         memory_size
@@ -53,9 +61,10 @@ fn check_int_real_memory(memory_size: usize) {
         println!("Current memory: {}", get_current_memory());
     }
     println!("End of check");
+    Ok(())
 }
 
-fn check_falt_real_memory(memory_size: usize) {
+fn check_falt(memory_size: usize) -> Result<(), ()> {
     println!(
         "Check falt memory overflow, ckb-vm memory size: {}",
         memory_size
@@ -72,9 +81,10 @@ fn check_falt_real_memory(memory_size: usize) {
         println!("Current memory: {}", get_current_memory());
     }
     println!("End of check");
+    Ok(())
 }
 
-fn check_asm_real_memory(memory_size: usize) {
+fn check_asm(memory_size: usize) -> Result<(), ()> {
     println!(
         "Check asm memory overflow, ckb-vm memory size: {}",
         memory_size
@@ -94,18 +104,68 @@ fn check_asm_real_memory(memory_size: usize) {
         println!("Current memory: {}", get_current_memory());
     }
     println!("End of check");
+    Ok(())
 }
 
-fn check_real_memory(memory_size: usize) {
-    check_int_real_memory(memory_size);
-    check_falt_real_memory(memory_size);
-    check_asm_real_memory(memory_size);
+fn check_asm_in_thread(memory_size: usize) -> Result<(), ()> {
+    println!(
+        "Check asm in thread memory overflow, ckb-vm memory size: {}",
+        memory_size
+    );
+    println!("Base memory: {}", get_current_memory());
+    for _ in 0..10 {
+        let asm_core = AsmCoreMachine::new(ISA_IMC, VERSION0, u64::max_value(), memory_size);
+        let core = DefaultMachineBuilder::new(asm_core).build();
+        let mut machine = AsmMachine::new(core);
+        machine
+            .load_program(&BIN_PATH_BUFFER, &vec![BIN_NAME.clone().into()])
+            .unwrap();
+        let thread_join_handle = thread::spawn(move || {
+            let result = machine.run();
+            assert!(result.is_ok());
+            assert_eq!(result.unwrap(), 0);
+        });
+        thread_join_handle.join().unwrap();
+        println!("Current memory: {}", get_current_memory());
+    }
+    println!("End of check");
+    Ok(())
+}
+
+fn test_memory(memory_size: usize) -> Result<(), ()> {
+    if check_int_real_memory(memory_size).is_err() {
+        return Err(());
+    }
+    if check_falt_real_memory(memory_size).is_err() {
+        return Err(());
+    }
+    if check_asm_real_memory(memory_size).is_err() {
+        return Err(());
+    }
+    if check_asm_in_thread(memory_size).is_err() {
+        return Err(());
+    }
+    Ok(())
 }
 
 fn main() {
+    epoch::advance().unwrap();
+    let base_allocated = stats::allocated::read().unwrap() as f64 * 1.02f64;
+    let base_resident = stats::resident::read().unwrap() as f64 * 1.02f64;
+
     let memory_size = 1024 * 1024 * 4;
-    check_real_memory(memory_size);
+    if test_memory(memory_size).is_err() {
+        panic!("run testcase failed");
+    }
+
+    assert!((stats::allocated::read().unwrap() as f64) < base_allocated);
+    assert!((stats::resident::read().unwrap() as f64) < base_resident);
 
     let memory_size = 1024 * 1024 * 2;
-    check_real_memory(memory_size);
+    if test_memory(memory_size).is_err() {
+        panic!("run testcase failed");
+    }
+
+    assert!((stats::allocated::read().unwrap() as f64) < base_allocated);
+    assert!((stats::resident::read().unwrap() as f64) < base_resident);
 }
