--- tests/tests_memory/Cargo.toml
@@ -1,12 +1,0 @@
-[package]
-name = "tests_memory"
-version = "0.1.0"
-edition = "2021"
-
-# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
-
-[dependencies]
-ckb-vm = { path= "../../", features = ["asm"] }
-jemallocator = "0.5.0"
-jemalloc-ctl = "0.5.0"
-lazy_static = "1.4.0"
