--- tests/tests_memory/src/main.rs
@@ -1,3 +1,0 @@
-pub mod misc;
-
-fn main() {}
